Create a new file to add mock multi_asic functions. Fix RFC4292 as per review comments.
