+++ src/sonic_ax_impl/mibs/ietf/rfc4292.py
@@ -58,8 +58,11 @@ class RouteUpdater(MIBUpdater):
         route_str = "ROUTE_TABLE:0.0.0.0/0"
 
         for db_conn in Namespace.get_non_host_dbs(self.db_conn):
-            # For multi-asic platform, proceed to get routes only for front end namespaces.
-            if front_ns and db_conn.namespace not in front_ns: continue
+            # For multi-asic platform, proceed to get routes only for 
+            # front end namespaces.
+            # For single-asic platform, front_ns will be empty list.
+            if front_ns and db_conn.namespace not in front_ns:
+                continue
             port_config = multi_asic.get_port_table(db_conn.namespace)
             ent = db_conn.get_all(mibs.APPL_DB, route_str, blocking=False)
             if ent is None:
@@ -70,12 +73,15 @@ class RouteUpdater(MIBUpdater):
                 ## Ignore non front panel interfaces
                 ## TODO: non front panel interfaces should not be in APPL_DB at very beginning
                 ## This is to workaround the bug in current sonic-swss implementation
-                if ifn == "eth0" or ifn == "lo" or ifn == "docker0": continue
+                if ifn == "eth0" or ifn == "lo" or ifn == "docker0":
+                    continue
                 # Ignore internal asic routes
-                if multi_asic.is_port_channel_internal(ifn, db_conn.namespace): continue
-                if ifn in port_config and \
-                   multi_asic.ROLE in port_config[ifn] and \
-                   port_config[ifn][multi_asic.ROLE] == multi_asic.INTERNAL_PORT : continue
+                if multi_asic.is_port_channel_internal(ifn, db_conn.namespace):
+                    continue
+                if (ifn in port_config and
+                   multi_asic.ROLE in port_config[ifn] and
+                   port_config[ifn][multi_asic.ROLE] == multi_asic.INTERNAL_PORT): 
+                    continue
                 sub_id = ip2tuple_v4(ipn.network_address) + ip2tuple_v4(ipn.netmask) + (self.tos,) + ip2tuple_v4(nh)
                 self.route_dest_list.append(sub_id)
                 self.route_dest_map[sub_id] = ipn.network_address.packed
