+++ tests/mock_tables/dbconnector.py
@@ -8,11 +8,9 @@ import redis
 import swsssdk
 from swsssdk import SonicV2Connector
 from swsssdk import SonicDBConfig
-from sonic_py_common import multi_asic
 from swsssdk.interface import DBInterface
 from swsscommon import swsscommon
 
-int_port_channel = ['PortChannel01', 'PortChannel02', 'PortChannel03', 'PortChannel04']
 
 if sys.version_info >= (3, 0):
     long = int
@@ -49,7 +47,9 @@ def load_database_config():
         sonic_db_file_path=os.path.join(
             os.path.dirname(os.path.abspath(__file__)), 'database_config.json'))
 
+
 _old_connect_SonicV2Connector = SonicV2Connector.connect
+
 
 def connect_SonicV2Connector(self, db_name, retry_on=True):
     ns_list = SonicDBConfig.get_ns_list()
@@ -141,52 +141,6 @@ class SwssSyncClient(mockredis.MockRedis):
         # Find every key that matches the pattern
         return [key for key in self.redis.keys() if regex.match(key)]
 
-def mock_get_num_asics():
-    ns_list = SonicDBConfig.get_ns_list()
-    if len(ns_list) > 1:
-        return(len(ns_list) - 1)
-    else:
-        return 1
-
-
-def mock_is_multi_asic():
-    if mock_get_num_asics() > 1:
-        return True
-    else:
-        return False
-
-def mock_get_all_namespaces():
-   if mock_get_num_asics() == 1:
-       return {'front_ns': [], 'back_ns': []}
-   else:
-       return {'front_ns': ['asic0', 'asic1'], 'back_ns': ['asic2']}
-
-def mock_is_port_channel_internal(port_channel, namespace=None):
-    if (mock_get_num_asics() == 1):
-        return False
-    else:
-        return True if port_channel in int_port_channel else False
-
-def mock_get_port_table(namespace=None):
-    if namespace is not None:
-        fname = os.path.join(INPUT_DIR, namespace, 'config_db.json')
-    else:
-        fname = os.path.join(INPUT_DIR, 'config_db.json')
-    port_table = {}
-    db = {}
-    with open(fname) as f:
-        db = json.load(f)
-    for k in db:
-        if 'PORT_TABLE' in db:
-            new_key = k[len('PORT_TABLE:'):]
-            port_table[new_key] = db[k]
-    return port_table
-        
-multi_asic.get_num_asics = mock_get_num_asics
-multi_asic.is_multi_asic = mock_is_multi_asic
-multi_asic.get_all_namespaces = mock_get_all_namespaces
-multi_asic.is_port_channel_internal = mock_is_port_channel_internal
-multi_asic.get_port_table = mock_get_port_table
 
 DBInterface._subscribe_keyspace_notification = _subscribe_keyspace_notification
 mockredis.MockRedis.config_set = config_set
